+++ fet/src/engine/studentsset.h
@@ -102,7 +102,7 @@ class StudentsSubgroup: public StudentsSet
 public:
 	int indexInInternalSubgroupsList;
 	
-	QList<qint16> activitiesForSubgroup;
+	QList<int> activitiesForSubgroup;
 
 	StudentsSubgroup();
 	~StudentsSubgroup();
